Update TRLToken.sol
the order listed in the CustomTaxPeriod must follow what is listed in the struct{} jus above it. If it does not, the taxes will not be allocated properly. In the case of the current contract setup, 4% will be allocated to marketing and 1% to devWallet

Marketing is 1% and Development is 4% for TRLToken